[feat] #154 ExtensionUtils.kt 생성
